Android agent: Create a class that hides the details of consrructing a POST to the WebpageTest server.
This forces the agent to encode information for upload in exactly the same way as other agents.  For example, a true value will always be the string '1'.  Minimizing differences in the POSTs agents make reduces the chances that a server change will break one agent and not another.

Change subrun to firstView in run class.

BUG=
TEST=

Review URL: https://codereview.appspot.com/6156043